rightbar responsive for mobile
