+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^1.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-loader-spinner": "^6.1.0",
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "socket.io-client": "^4.6.2",
