--- src/components/rightbar/rightbar.css
@@ -1,121 +1,0 @@
-.rightbarContainer{
-    flex: 3.5;
-}
-
-.rightbarWrapper{
-    padding: 20px 20px 0 0;
-}
-
-.birthdayContainer{
-    display: flex;
-    align-items: center;
-}
-
-.birthdayImg{
-    width: 45px;
-    height: 45px;
-    margin-right: 10px;
-}
-
-.birthdayText{
-    font-weight: 300;
-    font-size: 15px;
-}
-
-.adContainer{
-    margin: 30px 0;
-}
-
-.adImg{
-    width: 100%;
-    border-radius: 10px;
-    cursor: pointer;
-}
-
-.adTextContainer{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 5px;
-}
-
-.adText{
-    font-weight: 100;
-    text-align: center;
-    font-size: 15px;
-}
-
-.adIcon{
-    cursor: pointer;
-}
-
-h4{
-    margin-bottom: 10px;
-}
-
-.onlineFriendsList{
-    margin: 0;
-    padding: 0;
-    list-style: none;
-}
-
-.rightbarFollowButton{
-    margin-top: 30px;
-    margin-bottom: 10px;
-    border: none;
-    background-color: rgb(146, 239, 146);
-    color: white;
-    border-radius: 5px;
-    padding: 5px 10px;
-    display: flex;
-    align-items: center;
-    font-size: 16px;
-    font-weight: 500;
-    cursor: pointer;
-}
-
-.rightbarTitle{
-    font-size: 18px;
-    font-weight: 500;
-    margin-bottom: 10px;
-}
-
-.rightbarInfo{
-    margin-bottom: 30px;
-}
-
-.rightbarInfoItem{
-    margin-bottom: 10px;
-}
-
-.rightbarInfoKey{
-    font-weight: 500;
-    margin-right: 15px;
-    color: #555;
-}
-
-.rightbarInfoValue{
-    font-weight: 300;
-}
-
-
-.rightbarFollowings{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
-.rightbarFollowing{
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 20px;
-    cursor: pointer;
-}
-
-.rightbarFollowingImg{
-    height: 100px;
-    width: 100px;
-    object-fit: cover;
-    border-radius: 8px;
-    border: 2px solid rgb(146, 239, 146);
-}
