--- src/components/rightbar/rightbar.jsx
@@ -1,148 +1,0 @@
-import { Add, MoreVert, Remove } from "@mui/icons-material";
-import { Users } from "../../dummyData";
-import Online from "../online/online";
-import "./rightbar.css";
-import { useContext, useEffect, useState } from "react";
-import axios from "axios";
-import { Link } from "react-router-dom";
-import { AuthContext } from "../../context/authContext";
-
-const Rightbar = ({ profile, user }) => {
-  const PF = process.env.REACT_APP_PUBLIC_FOLDER;
-  const [friends, setFriends] = useState([]);
-  const { user: currentUser, dispatch } = useContext(AuthContext);
-  const [followed, setFollowed] = useState(
-    currentUser.following.includes(user?._id)
-  );
-
-  useEffect(() => {
-    setFollowed(currentUser.following.includes(user?._id));
-  }, [currentUser, user]);
-
-  useEffect(() => {
-    const getFriends = async () => {
-      try {
-        const friendList = await axios.get(`/users/following/${user?._id}`);
-        setFriends(friendList.data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    getFriends();
-  }, [user]);
-
-  const handleClick = async () => {
-    try {
-      if (followed) {
-        await axios.put(`/users/${user._id}/unfollow`, {
-          userId: currentUser._id,
-        });
-        dispatch({ type: "UNFOLLOW", payload: user._id });
-      } else {
-        await axios.put(`/users/${user._id}/follow`, {
-          userId: currentUser._id,
-        });
-        dispatch({ type: "FOLLOW", payload: user._id });
-      }
-    } catch (error) {
-      console.log(error);
-    }
-    setFollowed(!followed);
-  };
-
-  const HomeRightBar = () => {
-    return (
-      <div className="hidden md:block">
-        <div className="birthdayContainer">
-          <img src={PF + "birthday.jpg"} alt="" className="birthdayImg" />
-          <span className="birthdayText">
-            <b>Mr Moose</b> and <b>3 others</b> have their birthday today{" "}
-          </span>
-        </div>
-        <div className="adContainer">
-          <div className="adTextContainer">
-            <span className="adText">Sponsored Ad</span>
-            <MoreVert className="adIcon" />
-          </div>
-          <img src={PF + "ad.jpg"} alt="" className="adImg" />
-        </div>
-        <h4 className="rightbarOnlineFriends">Online Friends</h4>
-        <ul className="onlineFriendsList">
-          {Users.map((u) => (
-            <Online key={u.id} user={u} />
-          ))}
-        </ul>
-      </div>
-    );
-  };
-
-  const ProfileRightBar = () => {
-    return (
-      <>
-        {user.username !== currentUser.username && (
-          <button className="rightbarFollowButton" onClick={handleClick}>
-            {followed ? "Unfollow" : "Follow"}
-            {followed ? <Remove /> : <Add />}
-          </button>
-        )}
-        <h4 className="rightbarTitle">User Information</h4>
-        <div className="rightbarInfo">
-          <div className="rightbarInfoItem">
-            <span className="rightbarInfoKey">City: </span>
-            <span className="rightbarInfoValue">{user.city}</span>
-          </div>
-          <div className="rightbarInfoItem">
-            <span className="rightbarInfoKey">From: </span>
-            <span className="rightbarInfoValue">{user.from}</span>
-          </div>
-          <div className="rightbarInfoItem">
-            <span className="rightbarInfoKey">Relationship: </span>
-            <span className="rightbarInfoValue">
-              {user.relationship === 1
-                ? "Single"
-                : user.relationship === 2
-                ? "Married"
-                : "Complicated"}
-            </span>
-          </div>
-        </div>
-        <h4 className="userFriends">User Friends</h4>
-        <div className="rightbarFollowings">
-          {friends.map((friend) => (
-            <Link
-              to={`/profile/${friend.username}`}
-              style={{
-                textDecoration: "none",
-                color: "black",
-                textAlign: "center",
-              }}
-            >
-              <div className="rightbarFollowing">
-                <img
-                  src={
-                    friend.profilePicture
-                      ? PF + friend.profilePicture
-                      : PF + "person/10.png"
-                  }
-                  alt=""
-                  className="rightbarFollowingImg"
-                />
-                <span className="rightbarFollowingName">{friend.username}</span>
-              </div>
-            </Link>
-          ))}
-        </div>
-      </>
-    );
-  };
-
-  return (
-    <div className="rightbarContainer">
-      <div className="rightbarWrapper">
-        {profile ? <ProfileRightBar /> : <HomeRightBar />}
-      </div>
-    </div>
-  );
-};
-
-export default Rightbar;
