+++ src/pages/home.jsx
@@ -1,7 +1,7 @@
 // import "./home.css";
 import TopBar from "../components/topbar";
 import Leftbar from "../components/leftbar/leftbar";
-import Rightbar from "../components/rightbar/rightbar";
+import Rightbar from "../components/rightbar";
 import Feed from "../components/feed";
 
 const Home = () => {
